refactor(project): split projectSlice into domain reducer modules
Decompose the 673-LOC projectSlice god-file into 12 focused reducer
modules under features/project/reducers/ (meta, character, world,
outline, manuscript, relationship, plot, writingAnalytics, binder,
storyObject, mindMap, interview). Each module exports plain case-reducer
objects typed against a new shared ProjectSliceState; the slice spreads
them into one createSlice.

The single redux-undo wrapper (app/store.ts), the `project/*` action
namespace, the ProjectData shape, extraReducers, and the
streamInterviewChunk matcher are all unchanged — pure file
decomposition, zero behavior change. initialState moves to projectState
to avoid a circular import from metaReducers.resetProject.

Verified: typecheck, lint (0 warnings), and all 76 projectSlice +
projectSelectors parity tests pass.
